Update tests for profile settings
