+++ spec/features/profiles/user_edit_profile_spec.rb
@@ -7,7 +7,7 @@ RSpec.describe 'User edit profile' do
 
   before do
     sign_in(user)
-    visit(edit_user_registration_path)
+    visit(profile_path)
   end
 
   it_behaves_like 'validates username' do
@@ -21,7 +21,7 @@ RSpec.describe 'User edit profile' do
         page.find('body').click
         wait_for_requests
 
-        expect(find('.user_username')).not_to have_css '.is-invalid'
+        expect(find('#user_username')[:class]).not_to include 'is-invalid'
       end
     end
 
@@ -46,52 +46,18 @@ RSpec.describe 'User edit profile' do
 
         click_button 'btn_edit_profile'
 
-        expect(page).to have_content(t('simple_form.error_notification.default_message'))
         expect(page).to have_content(taken_message_for_email)
-      end
-    end
-
-    context 'when change password without current password', :js do
-      before do
-        fill_in 'user_password',              with: user.password
-        fill_in 'user_password_confirmation', with: user.password
-      end
-
-      it 'shows current password is required message' do
-        click_button 'btn_edit_profile'
-
-        expect(page).not_to have_content(t('simple_form.error_notification.default_message'))
-        expect(page).to have_content(blank_message_for_current_password)
-      end
-
-      it 'does not reload the page' do
-        expect_page_to_not_reload do
-          click_button 'btn_edit_profile'
-        end
-      end
-    end
-
-    context 'with invalid current password' do
-      it 'shows current password is invalid message' do
-        fill_in 'user_password',              with: user.password
-        fill_in 'user_password_confirmation', with: user.password
-        fill_in 'user_current_password',      with: 'invalid-password'
-
-        click_button 'btn_edit_profile'
-
-        expect(page).to have_content(t('simple_form.error_notification.default_message'))
-        expect(page).to have_content(invalid_message_for_current_password)
       end
     end
   end
 
   context 'with no errors' do
-    context 'without updating email and password' do
+    context 'without updating anything' do
       it 'updates profile' do
         click_button 'btn_edit_profile'
 
         expect(page).to have_content(t('devise.registrations.updated'))
-        expect(page).to have_current_path(edit_user_registration_path(user))
+        expect(page).to have_current_path(profile_path)
       end
     end
 
@@ -109,7 +75,7 @@ RSpec.describe 'User edit profile' do
 
         expect(current_email.subject).to eq t('devise.mailer.email_changed.subject')
         expect(current_email.to.first).to eq user.email
-        expect(page).to have_content(t('devise.registrations.update_needs_confirmation'))
+        expect(page).to have_content(t('devise.registrations.updated'))
       end
 
       it 'sends confirmation instructions to new email' do
@@ -119,43 +85,21 @@ RSpec.describe 'User edit profile' do
         expect(current_email.subject).to eq t('devise.mailer.confirmation_instructions.subject')
       end
     end
-
-    context 'when password is updated with valid current password' do
-      it 'sends password changed notification' do
-        fill_in 'user_password',              with: user.password
-        fill_in 'user_password_confirmation', with: user.password
-        fill_in 'user_current_password',      with: user.password
-
-        click_button 'btn_edit_profile'
-
-        open_email(user.email)
-
-        expect(current_email.subject).to eq t('devise.mailer.password_change.subject')
-        expect(page).to have_content(t('devise.registrations.updated'))
-      end
-    end
   end
 
+  # TODO: cancel account
   # rubocop:disable RSpec/PredicateMatcher
-  describe 'when I cancel my account', :js do
-    it 'deletes user' do
-      accept_alert do
-        click_button 'btn_cancel_my_account'
-      end
+  # describe 'when I cancel my account', :js do
+  #   it 'deletes user' do
+  #     accept_alert do
+  #       click_button 'btn_cancel_my_account'
+  #     end
 
-      expect(page).to have_current_path root_path
-      expect(page).to have_content(t('devise.registrations.destroyed'))
-      expect(User.exists?(user.id)).to be_falsy
-    end
-  end
+  #     expect(page).to have_current_path root_path
+  #     expect(page).to have_content(t('devise.registrations.destroyed'))
+  #     expect(User.exists?(user.id)).to be_falsy
+  #   end
+  # end
   # rubocop:enable RSpec/PredicateMatcher
-
-  def blank_message_for_current_password
-    t('activerecord.errors.models.user.attributes.current_password.blank')
-  end
-
-  def invalid_message_for_current_password
-    t('activerecord.errors.models.user.attributes.current_password.invalid')
-  end
 end
 # rubocop:enable RSpec/DescribeClass
